+++ Tests/DeserializerTests/TOMLInvalidationTests.swift
@@ -629,4 +629,10 @@ final class TOMLInvalidationTests: XCTestCase {
         let tomlData = try Data(contentsOf: tomlURL)
         XCTAssertThrowsError(try TOMLDeserializer.tomlTable(with: tomlData))
     }
+
+    func test_array_extending_table() throws {
+        let tomlURL = URL(fileURLWithPath: "\(self.directory)\(kSeparator)array-extending-table.toml", isDirectory: false)
+        let tomlData = try Data(contentsOf: tomlURL)
+        XCTAssertThrowsError(try TOMLDeserializer.tomlTable(with: tomlData))
+    }
 }